Fixing some warning and added general secret 1-1 structure

+++ 05-SceneManager/Game.cpp
@@ -37,7 +37,7 @@ void CGame::TogglePause()
 	
 }
 
-void CGame::PauseFor(DWORD milliseconds)
+void CGame::PauseFor(ULONGLONG milliseconds)
 {
 	if (!isPaused)
 	{
